requirements.txt
streamlit
yfinance
pandas
groq
python-dotenv